\name{CondB}
\alias{CondB}
\docType{data}
\title{Stress condition for animals contained in the data set Nuclei
}
\description{
Animals measured in the Nuclei data set were either subjected to chronic restraint stress (stress) or not (control).  
}
\usage{
data(CondB)
}

\format{
  A factor vector indicating which stress group the individuals in Nuclei belong to.
}
%\details{
%
%}
\source{
The data are provided courtesy of David Crews at the University of Texas at Austin.
}
\references{
Crews, D, R Gillette, SV Scarpino, M Manikkam, MI Savenkova, MK Skinner. 2012.
Epigenetic Transgenerational Alterations to Stress Response in Brain Gene Networks and 
Behavior. Proc. Natl. Acad. Sci. USA. 109 (23). 9143 - 9148.
}

\examples{
data(CondB)
}
