\name{Dyad}
\docType{data}
\alias{Dyad}

\title{Housing dyad for animals contained in the data set Nuclei
}

\description{
Animals measured in the Nuclei data set were housed in dyads with one individual from the Vinclozolin line and one from the control line housed together.  Each dyad was either stressed or not stressed.    
}

\usage{
data(Dyad)
}

\format{
  A factor vector indicating which housing dyad the individuals in Nuclei are in.
}

%\details{
%
%}

\source{
The data are provided courtesy of David Crews at the University of Texas at Austin.
}

\references{
Crews, D, R Gillette, SV Scarpino, M Manikkam, MI Savenkova, MK Skinner. 2012.
Epigenetic Transgenerational Alterations to Stress Response in Brain Gene Networks and 
Behavior. Proc. Natl. Acad. Sci. USA. 109 (23). 9143 - 9148.
}

\examples{
data(Dyad)
}
