\name{Groups}

\alias{Groups}
\docType{data}
\title{The group ID for animals contained in the data set Nuclei
}

\description{
Animals measured in the Nuclei data set belong to one of four groups determined by their linneage (Vinclozolin or Control) and their stress treatment (Stressed or Non-Stressed). 
}

\usage{
data(Groups)
}

\format{
  A factor vector indicating which group the individuals in Nuclei are in.
}

%\details{
%
%}

\source{
The data are provided courtesy of David Crews at the University of Texas at Austin.
}

\references{
Crews, D, R Gillette, SV Scarpino, M Manikkam, MI Savenkova, MK Skinner. 2012.
Epigenetic Transgenerational Alterations to Stress Response in Brain Gene Networks and 
Behavior. Proc. Natl. Acad. Sci. USA. 109 (23). 9143 - 9148.
}

\examples{
data(Groups)
}
