\name{plotBinomPower}
\alias{plotBinomPower}
\title{A function to plot the results of a binomPower run
}
\description{
A function to plot the results of a binomPower run.
}
\usage{
plotBinomPower(datPlotBig,params)
}
\arguments{
  \item{datPlotBig}{
a (non-empty) matrix of data values, with columns trueTau, ndads, trueTau2
}
  \item{params}{
a (non-empty) matrix of parameter values, with columns mm and vv. 
}
}
\value{
Returns a list of two plots of the binomPower analysis results.
}
\examples{
#not run
}

